requirements.txt
kivy>=2.3.0
PySocks>=1.7.1
paramiko>=3.4.0
keyring>=25.0.0
Pillow>=10.0.0
pytest>=8.0.0
pytest-mock>=3.14.0
